program description
The purpose of this program is meant to replicate the functions of an operating system. More specifically, the program provides the user with 4 options. The first option checks whether there are enough memory blocks in the memory block table (MBT) before initiating a process. Once the program determines that there are enough blocks, a process control block (PCB) is made. This PCB contains information that is unique to that process, such as a process ID (PID), page table size, and a pointer to the page table. The data structure is then saved into a linked list. Option 2 allows the user to see the PID, the number of blocks, and the specific blocks allocated to each PCB in the list. Option 3 gives the user the option to delete a PCB from the list by specifying its PID. The program looks to see if a PCB with that PID exists. If it does, the PCB is deleted along with the page table and the memory blocks are returned to the MBT. If the process does not exist, an error message is displayed. Option 4 is used to exit the program. However, if the user chooses to exit while there are processes in the list, a list of the process PIDs is displayed. The user is then asked to confirm that they want to exit the program. If they proceed, the memory that was dynamically allocated for each PCB is deleted by calling the PCB destructor for each PCB in the linked list. The program is made up of a struct for PCB and MBT. Additionally, there also exists a linked list class, which is used to hold all the PCBs. The page table pointer for each PCB points to an integer array that is dynamically allocated using the total size of the space address for each process. This number varies within the range of 10-250 blocks.