IPC description
The purpose of this program is to show an implementation of interprocess communication using a message queue. The program starts of by creating two child processes from main. Before the creation of these child processes, a message queue is created in main() with specific restrictions on what actions can be performed on the queue and who can take those actions. The first child process is created subsequently. In this child process, there is a request made to read a specific type of message (113) from the message queue. Because there still hasn’t been anything placed in the queue of that type, the process is suspended by the last parameter in the msgrcv() system call. The second child process begins. This process creates a message of type 12, which is supposed to never be read, and writes it to the queue. Then, child process 2 displays a greeting message before writing to the queue once again. The new message is of type 113. After this message is saved in the queue, a message of type 113 is now available. The child process 2 now attempts to receive a message type that doesn’t exist in the queue, so it is suspended. The message type that the second child process made before being suspended was requested earlier by child process 1, so it resumes execution. Process child 1 displays a message acknowledging having received the message. The first child process prepares a message and then saves it to the queue as type 114, which is the type that child process 2 requested. Once the message is saved, child process 1 terminates. Since there now exists a message of type 114 in the queue, child process 2 can resume. The process then confirms it has received the message by displaying a message. The second child process ends after. After both child processes have ended, the message queue is removed back in main(). The program ends by displaying a message that the parent process, main(), has ended.