IPC using a shared memory segment
The lab demonstrates how the other form of interprocess communication works. The shmp1 program is launched initially. It begins by calling the shm_init() function to create the shared memory segment. It also attaches the shared memory segment to its own address space. This function also performs the task of pointing to the shared memory pointer to the struct “myclass”. Within this structure is where information that is to be used by all other processes sharing the segment lies. The shm_init() method returns the identifier of the segment. Afterwards, the parent process goes on to spawn 3 children. Initially, they share the same address space as the master process. However, the execl() system call causes the children to replace their parent’s address space with a new program, shmc1. Each one of the 3 children receives the identifier of the shared memory segment and a unique name. After the parent has created the children, it is made to wait for them to terminate before moving forward. This occurs in wait_and_wrap_up(). The children execute the shmc1 program. This program stores the information that the parent passed to the children, such as the unique name and the identifier of the shared memory segment. Each of the 3 child processes attach the shared memory segments to their address space. Then they each execute the sell_seats() function. Inside this function they attempt to sell the number of seats, which was originally 15 in the shared segment. Before selling a seat  there exists a sleep() function that puts the program to sleep. The amount of time the program sleeps for is influenced by the process ID of the process itself. It is also a value that is random. Since the processes run concurrently, this means that a child process could be sleeping while the other two have each sold a seat. Likewise, there is another call to sleep() before printing the current amount of available seats. This means that at any time before decrementing the seat count or printing the current seat count, the value could have been tampered by the other processes. The output is different each time you run shmp1 because we will never know which child will finish first. After the seats have been sold completely, the child processes return from the sell_seats() function and detach the shared memory segment from their address space. Each child then terminates. At this point the parent is able to resume. The parent itself detaches the segment and the calls shmctl() to delete it. In order for this to happen, the value, shm_nattch, contained within the shmid_ds structure must be 0 (indicating that there are no more processes attaching the segment). The program ends.